+++ apps/overlay/src/views/lowerthirds/LowerThirds.o.tsx
@@ -2,8 +2,11 @@ import { Box, Container } from "@mantine/core"
 import { LowerthirdData } from "interface/ws"
 import { useLt } from "utils/hooks"
 import LowerTicker from "./LowerTicker.o"
+import { ReactNode } from "react"
 
-const switchLT = (mode: any) => {
+type SwitchLT = (mode: keyof LowerthirdData) => ReactNode
+
+const switchLT: SwitchLT = (mode) => {
   switch (mode) {
     case "ad":
       return <></>
@@ -13,6 +16,8 @@ const switchLT = (mode: any) => {
       return <></>
     case "ticker":
       return <LowerTicker />
+    default:
+      return null
   }
 }
